Fix some signedness

+++ views/GBScoresView.cpp
@@ -168,8 +168,8 @@ void GBScoresView::DrawScores(const GBScores & scores, bool allRounds) {
 		DrawStringLongPair("Doubletime:", doubletime, col, 80, 9, GBColor::black);
 //other
 	if ( scores.Population() ) {
-		DrawStringPair("Economy:", ToPercentString(scores.EconFraction(), 0), col, 90, 9, GBColor::black);
-		DrawStringPair("Combat:", ToPercentString(scores.CombatFraction(), 0), col, 100, 9, GBColor::black);
+		DrawStringPair("Economy:", ToPercentString((float)scores.EconFraction(), 0), col, 90, 9, GBColor::black);
+		DrawStringPair("Combat:", ToPercentString((float)scores.CombatFraction(), 0), col, 100, 9, GBColor::black);
 	}
 	if ( side )
 		DrawStringLongPair("Territory:", scores.Territory(), col, 110, 9, GBColor::black);
